toasty/pipeline/djangoplicity.py: fix incoming JSON
Some Djangoplicity versions/installs have an issue where their JSON API
services should be returning Unicode text, but the data come through
text containing reprs of Python `bytes` objects. This is definitely a
bug, but fortunately it's not too hard for us to patch up the output
to turn it into what it should be.